Matchmaking: submatchmaking deleted and waiting game updated. Now if the subscription cannot connect to the server, an error message is displayed to the user
